refactor: extract shared data model utilities into ied-data-model.ts
Create new utils/ directory.
